Struct merkle_tree::MerkleTree

pub struct MerkleTree<T: ToString + Display + Clone> { /* fields omitted */ }

Methods

impl<T: ToString + Display + Clone> MerkleTree<T>

fn new() -> Self

Creates new, empty MerkleTree.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let tree = MerkleTree::new();
assert_eq!(0, tree.len());

fn from_vec(data: Vec<T>) -> Self

Creates MerkleTree from Vec of elements.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let values = vec![1, 2, 3, 4];
let tree = MerkleTree::from_vec(values);
assert_eq!(4, tree.len());

fn push(&mut self, value: T)

Push element into the end of the tree.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
tree.push(1);
assert_eq!(1, tree.len());

fn remove(&mut self, index: usize) -> bool

Removes element from the tree and returns true if element was removed successfully and false if index out of bounds.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::from_vec(vec![1, 2, 3]);
assert!(tree.remove(1));
assert_eq!(2, tree.len());
assert!(!tree.remove(5));

fn get(&self, index: usize) -> Option<&T>

Retrieves an element in the MerkleTree by index.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
tree.push(1);
tree.push(2);
tree.push(3);
assert_eq!(tree.get(1), Some(&2));

fn get_values(&self) -> Option<Vec<T>>

Retrieves copies of all elements in the MerkleTree.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
tree.push(1);
tree.push(2);
tree.push(3);
assert_eq!(tree.get_values(), Some(vec![1, 2, 3]));

fn len(&self) -> usize

Returns the number of elements in the three

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
assert_eq!(tree.len(), 0);
tree.push(1);
assert_eq!(tree.len(), 1);

fn height(&self) -> usize

Returns the height of the three

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
assert_eq!(tree.height(), 0);
tree.push(1);
assert_eq!(tree.height(), 0);
tree.push(2);
assert_eq!(tree.height(), 1);
tree.push(3);
assert_eq!(tree.height(), 2);

fn is_empty(&self) -> bool

Returns true if the MerkleTree is empty.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
assert!(tree.is_empty());
tree.push(1);
assert!(!tree.is_empty());

fn root_hash(&self) -> Option<&String>

Returns root hash of MerkleTree

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::new();
assert_eq!(Some(&"5feceb66ffc86f38d952786c6d696c79c2dbc239dd4e91b46729d73a27fb57e9".to_string()), tree.root_hash());
tree.push(1);
tree.push(2);
tree.push(3);
tree.push(4);
assert_eq!(Some(&"85df8945419d2b5038f7ac83ec1ec6b8267c40fdb3b1e56ff62f6676eb855e70".to_string()), tree.root_hash());

fn validate_element(&self, value: T, root_hash: &str) -> bool

Verifies if the value really presents in MerkleTree

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let mut tree = MerkleTree::from_vec(vec![1, 2, 3, 4]);
assert!(tree.validate_element(2, "85df8945419d2b5038f7ac83ec1ec6b8267c40fdb3b1e56ff62f6676eb855e70");

fn iter(&self) -> Iter<Rc<T>>

Returns a front-to-back iterator.

Examples

extern crate merkle_tree;
use merkle_tree::MerkleTree;

let tree = MerkleTree::from_vec(vec![1, 2, 3, 4]);
let vec: Vec<Rc<i32>> = tree.iter().collect();
assert_eq!(4, vec.len());